+++ packages/theme_tailor/lib/src/template/theme_class_template.dart
@@ -47,7 +47,7 @@ class ThemeClassTemplate {
     if (config.themes.isEmpty) return '';
     final buffer = StringBuffer();
     config.themes.forEachIndexed((i, e) {
-      buffer.write(_themeTemplate(i, e, config.fields.keys.toList()));
+      buffer.write(_themeTemplate(i, e));
     });
     final themesList = config.themes.fold('', (p, theme) => '$p$theme,');
     buffer.writeln('static final ${config.themesFieldName} = [$themesList];');
@@ -55,16 +55,24 @@ class ThemeClassTemplate {
   }
 
   /// Template for one static theme
-  String _themeTemplate(int index, String themeName, List<String> props) {
+  String _themeTemplate(int index, String themeName) {
     final buffer = StringBuffer();
     final returnType = config.className;
 
-    for (final prop in props) {
-      buffer.write('$prop: ${config.baseClassName}.$prop[$index],');
+    for (final field in config.fields.entries) {
+      final values = field.value.values;
+      if (values != null) {
+        buffer.write('${field.key}: ${values[index]},');
+      } else {
+        buffer.write(
+            '${field.key}: ${config.baseClassName}.${field.key}[$index],');
+      }
     }
 
+    final themeModifier = config.constantThemes ? 'const' : 'final';
+
     return '''
-    static final $returnType $themeName = $returnType(
+    static $themeModifier $returnType $themeName = $returnType(
       ${buffer.toString()}
     );\n
     ''';